updated documentation, setup.py
also matplotlib added as dependency, moved functions from ArucoFunc to ArucoHelper because it made more sense. Other cleanup like variables names
